Moved the Config.java code into its own Maven build.
- Moved the Config.java code into its own Maven directory structure.
- Updated the aggregate POM file.
- Updated the references to Config.java to reflect this change.
- The "Common" Maven build should contain code that can be used by
  other tools that we have created and not just for FactorBase.
