+++ factorbase/pom.xml
@@ -63,11 +63,6 @@
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>mysql</groupId>
-      <artifactId>mysql-connector-java</artifactId>
-      <version>5.1.22</version>
-    </dependency>
-    <dependency>
       <groupId>gov.nist.math</groupId>
       <artifactId>jama</artifactId>
       <version>1.0.3</version>
@@ -96,6 +91,11 @@
       <groupId>colt</groupId>
       <artifactId>colt</artifactId>
       <version>1.2.0</version>
+    </dependency>
+    <dependency>
+      <groupId>ca.sfu.cs</groupId>
+      <artifactId>common</artifactId>
+      <version>1.0-SNAPSHOT</version>
     </dependency>
   </dependencies>
 </project>
