+++ factorbase/src/main/java/ca/sfu/cs/factorbase/exporter/csvexporter/CSVPrecomputor.java
@@ -15,7 +15,7 @@ import javax.swing.JOptionPane;
 
 import org.apache.commons.lang.StringUtils;
 
-import ca.sfu.cs.factorbase.app.Config;
+import ca.sfu.cs.common.Configuration.Config;
 
 import com.mysql.jdbc.Connection;
 /* zqian@Nov 21, fixed on bug for loading data into csv files (have to close the file)
