+++ factorbase/src/main/java/ca/sfu/cs/factorbase/tables/CPGenerator.java
@@ -1,6 +1,6 @@
 package ca.sfu.cs.factorbase.tables;
 
-import ca.sfu.cs.factorbase.app.Config;
+import ca.sfu.cs.common.Configuration.Config;
 
 import com.mysql.jdbc.Connection;
