+++ factorbase/src/main/java/ca/sfu/cs/factorbase/tables/KLD_generator.java
@@ -38,7 +38,7 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Set;
 
-import ca.sfu.cs.factorbase.app.Config;
+import ca.sfu.cs.common.Configuration.Config;
 import ca.sfu.cs.factorbase.util.QueryGenerator;
