+++ pom.xml
@@ -8,6 +8,7 @@
   <name>factorbase-aggregator</name>
   <url>https://github.com/sfu-cl-lab/FactorBase</url>
   <modules>
+    <module>common</module>
     <module>factorbase</module>
     <module>mlnexporter</module>
   </modules>
